MH3-116882: Regenerate apps

+++ apps/prime_apps/prime_base_1_4_modem/src/config/pic32cxmtg_pl460_rf215/stack/pal/pal_rf.c
@@ -214,7 +214,7 @@ static void lPAL_RF_DataCfmCb(DRV_RF215_TX_HANDLE txHandle,
     {
         /* Error confirm, called from PAL_RF_Tasks */
         dataCfm.bufId = (uint8_t)ctxt;
-        dataCfm.txTime = pCfmObj->timeIniCount;
+        dataCfm.txTime = (uint32_t)pCfmObj->timeIniCount;
     }
     else
     {
